Add deletion of demo services
