+++ demo/dashboard/server/serverless.yml
@@ -1,4 +1,4 @@
-service: demo-dash-server-${opt:username}
+service: demo-dash-server-${self:custom.EUBFR_USERNAME}
 
 plugins:
   - serverless-stack-output
@@ -8,8 +8,10 @@ plugins:
 custom:
   webpackIncludeModules: true
   eubfrEnvironment: ${opt:eubfr_env, file(../../../config.json):eubfr_env, env:EUBFR_ENV, 'dev'}
-  PRODUCER_KEY_ID: ${opt:PRODUCER_KEY_ID, file(../../../config.json):demo.${opt:username}.AWS_ACCESS_KEY_ID, env:PRODUCER_KEY_ID, ''}
-  PRODUCER_SECRET_ACCESS_KEY: ${opt:PRODUCER_SECRET_ACCESS_KEY, file(../../../config.json):demo.${opt:username}.AWS_SECRET_ACCESS_KEY, env:PRODUCER_SECRET_ACCESS_KEY, ''}
+  # EUBFR_USERNAME comes from cli parameter, but for spawn() and similar, it's from env var.
+  EUBFR_USERNAME: ${opt:username, file(../../../config.json):username, env:EUBFR_USERNAME, ''}
+  PRODUCER_KEY_ID: ${opt:PRODUCER_KEY_ID, file(../../../config.json):demo.${self:custom.EUBFR_USERNAME}.AWS_ACCESS_KEY_ID, env:PRODUCER_KEY_ID, ''}
+  PRODUCER_SECRET_ACCESS_KEY: ${opt:PRODUCER_SECRET_ACCESS_KEY, file(../../../config.json):demo.${self:custom.EUBFR_USERNAME}.AWS_SECRET_ACCESS_KEY, env:PRODUCER_SECRET_ACCESS_KEY, ''}
   SIGNED_UPLOADS_API: ${opt:SIGNED_UPLOADS_API, file(../../../services/storage/signed-uploads/.serverless/stack-output.json):ServiceEndpoint, env:SIGNED_UPLOADS_API, ''}
   DELETER_API: ${opt:DELETER_API, file(../../../services/storage/deleter/.serverless/stack-output.json):ServiceEndpoint, env:DELETER_API, ''}
   META_INDEX_API: ${opt:META_INDEX_API, file(../../../services/storage/meta-index/.serverless/stack-output.json):ServiceEndpoint, env:META_INDEX_API, ''}
