+++ scripts/delete.js
@@ -1,14 +1,8 @@
 #!/usr/bin/env node
 
 // Dependencies
-const { spawn } = require('child_process');
-const fs = require('fs');
-const path = require('path');
+const deleteServerlessService = require('./utils/deleteServerlessService');
 
-// Get config
-const config = require(`../config.json`);
-
-// Easier to edit in the beginning of the file
 const services = [
   `storage-objects`,
   `ingestion-etl-results`,
@@ -25,36 +19,4 @@ const services = [
   `value-store-projects`,
 ];
 
-// Helpers
-const resolveSymbolicLink = filePath => {
-  const lstat = fs.lstatSync(filePath);
-  return lstat.isSymbolicLink()
-    ? path.resolve(path.dirname(filePath), fs.readlinkSync(filePath))
-    : false;
-};
-
-const deleteServerlessService = service => {
-  console.log(`Start deletion of ${service}`);
-  console.time(service);
-
-  const operation = spawn(
-    `./node_modules/.bin/sls`,
-    ['remove', `--stage ${config.stage}`, `--region ${config.region}`],
-    {
-      cwd: resolveSymbolicLink(`node_modules/@eubfr/${service}`),
-    }
-  );
-
-  operation.stdout.on('data', data => {
-    console.log('\x1b[36m', `${data}`, '\x1b[0m');
-  });
-
-  operation.stderr.on('data', data => {
-    console.log('\x1b[31m', `${data}`, '\x1b[31m');
-  });
-
-  operation.on('close', () => console.timeEnd(service));
-};
-
-// And run the deletion procedure
 services.forEach(deleteServerlessService);
